+++ lib/src/plugin.web.dart
@@ -11,7 +11,7 @@ extension type VideoViewPlugin._(JSObject _) implements JSObject {
     _, {
     Object? params,
   }) {
-    const requiredVersion = '1.2.3';
+    const requiredVersion = '1.2.4';
     if (version != requiredVersion) {
       Zone.current.handleUncaughtError(
         'VideoViewPlugin.js version: $version. Required: $requiredVersion.\nPlease try cleaning the browser cache or using "dart run video_view:webinit" to update.',
@@ -41,6 +41,7 @@ extension type VideoViewPlugin._(JSObject _) implements JSObject {
   external void setPreferredAudioLanguage(String language);
   external void setPreferredSubtitleLanguage(String language);
   external void setShowSubtitle(bool show);
+  external void setKeepScreenOn(bool keepOn);
   external void setOverrideAudio(String? trackId);
   external void setOverrideSubtitle(String? trackId);
   external bool setDisplayMode(int displayMode);